Add hook for code mirror option generation
